Types and errors split in sampling behaviour
