+++ tests/src/nodes/validator.rs
@@ -21,7 +21,7 @@ use nomos_da_indexer::{
     IndexerSettings,
 };
 use nomos_da_network_core::{
-    protocols::sampling::behaviour::SubnetsConfig,
+    protocols::sampling::SubnetsConfig,
     swarm::{BalancerStats, DAConnectionPolicySettings, MonitorStats},
 };
 use nomos_da_network_service::{
